requirements.txt
anyascii==0.3.2
arrow==1.3.0
asgiref==3.7.2
beautifulsoup4==4.11.2
binaryornot==0.4.4
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
cookiecutter==2.4.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.6
django-filter==23.3
django-modelcluster==6.1
django-permissionedforms==0.1
django-rest-knox==4.2.0
django-taggit==4.0.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
drf-yasg==1.21.7
et-xmlfile==1.1.0
filetype==1.2.0
firebase-admin==6.2.0
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
l18n==2021.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgpack==1.0.7
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
pillow-heif==0.13.1
proto-plus==1.22.3
protobuf==4.24.4
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python-webpack-boilerplate==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
rich==13.6.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
text-unidecode==1.3
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
wagtail==5.1.2
webencodings==0.5.1
Willow==1.6.2
